refactor: Separated model import and replace default materials
