+++ scripts/setup_wizard/genshin_import_character_model.py
@@ -16,28 +16,11 @@ try:
 except Exception:
     print('Error! Run the first step of setup_wizard! Need to set up python script paths')
 
-material_assignment_mapping = {
-    'Yelan': {
-        'miHoYo - Genshin Dress1': 'Body',
-        'miHoYo - Genshin Dress2': 'Hair'
-    },
-    'Collei': {
-        'miHoYo - Genshin Dress': 'Hair'
-    },
-    'Ganyu': {
-        'miHoYo - Genshin Dress': 'Body'
-    },
-    'Rosaria': {
-        'miHoYo - Genshin Dress1': 'Hair',
-        'miHoYo - Genshin Dress2': 'Body'
-    }
-}
-
 
 class GI_OT_GenshinImportModel(Operator, ImportHelper):
     """Select the folder with the desired model to import"""
     bl_idname = "file.genshin_import_model"  # important since its how we chain file dialogs
-    bl_label = "3_Genshin: Select Model Folder"
+    bl_label = "2_Genshin: Import Character Model - Select Character Model Folder"
 
     # ImportHelper mixin class uses this
     filename_ext = "*.*"
@@ -61,17 +44,8 @@ class GI_OT_GenshinImportModel(Operator, ImportHelper):
     def execute(self, context):
         character_model_folder_file_path = self.file_directory if self.file_directory else os.path.dirname(self.filepath)
 
-        character_name = ''
-        for name, folder, files in os.walk(character_model_folder_file_path):
-            for file in files:
-                for tmp_character_name in material_assignment_mapping.keys():
-                    if tmp_character_name in file:
-                        character_name = tmp_character_name
-
-        self.__import_character_model(character_model_folder_file_path)
-        self.__replace_default_materials_with_genshin_materials(character_name)
-
         print(character_model_folder_file_path)
+        self.__import_character_model(character_model_folder_file_path)
 
         invoke_next_step(self.next_step_idx, character_model_folder_file_path)
         return {'FINISHED'}
@@ -86,56 +60,6 @@ class GI_OT_GenshinImportModel(Operator, ImportHelper):
             for file_name in files:
                 if '.fbx' in pathlib.Path(file_name).suffix:
                     return os.path.join(root, file_name)
-    
-    def __replace_default_materials_with_genshin_materials(self, character_name):
-        # TODO: We actually are specifically looking at meshes here...impacts unjoined workflows
-        body_mesh_object = [object for object in bpy.context.scene.objects if object.type == 'MESH'][0]
-
-        meshes = [mesh for mesh in bpy.context.scene.objects if mesh.type == 'MESH']
-
-        for mesh in meshes:
-            for material_slot in mesh.material_slots:
-                material_name = material_slot.name
-                mesh_body_part_name = material_name.split('_')[-1]
-                genshin_material = bpy.data.materials.get(f'miHoYo - Genshin {mesh_body_part_name}')
-                print(genshin_material)
-
-                if genshin_material:            
-                    material_slot.material = genshin_material
-                elif 'Dress' in mesh_body_part_name:
-                    print('Dress detected on character model!')
-                    material_mapping = material_assignment_mapping.get(character_name)
-                    print(material_mapping)
-
-                    if material_mapping:
-                        body_part = material_mapping.get(f'miHoYo - Genshin {mesh_body_part_name}')
-                        genshin_material = self.__clone_material_and_rename(material_slot, f'miHoYo - Genshin {body_part}', mesh_body_part_name)
-                    else:
-                        genshin_material = self.__clone_material_and_rename(material_slot, f'miHoYo - Genshin Body', mesh_body_part_name)
-                elif material_name == 'miHoYoDiffuse':
-                    material_slot.material = bpy.data.materials.get(f'miHoYo - Genshin Body')
-                    continue
-                else:
-                    print(f'Ignoring unknown mesh body part in character model: {mesh_body_part_name}')
-                    continue
-
-                # Don't need to duplicate multiple Face shader nodes
-                if genshin_material.name != f'miHoYo - Genshin Face':
-                    genshin_main_shader_node = genshin_material.node_tree.nodes.get('Group.001')
-                    genshin_main_shader_node.node_tree = self.__clone_shader_node_and_rename(genshin_material, mesh_body_part_name)
-        print('Replaced default materials with Genshin shader materials...')
-    
-    def __clone_material_and_rename(self, material_slot, mesh_body_part_name_template, mesh_body_part_name):
-        new_material = bpy.data.materials.get(mesh_body_part_name_template).copy()
-        new_material.name = f'miHoYo - Genshin {mesh_body_part_name}'
-
-        material_slot.material = new_material
-        return new_material
-
-    def __clone_shader_node_and_rename(self, material, mesh_body_part_name):
-        new_shader_node_tree = material.node_tree.nodes.get('Group.001').node_tree.copy()
-        new_shader_node_tree.name = f'miHoYo - Genshin {mesh_body_part_name}'
-        return new_shader_node_tree
 
 
 def register():
